+++ lib/config.ts
@@ -10,10 +10,14 @@ export const SUBSCRIPTION_URL = "https://www.patreon.com/abdibrokhim/membership"
 
 export const EMAIL = "abdibrokhim@gmail.com";
 
-export const APP_VERSION = "0.0.0"; // Current app version
+export const APP_VERSION = "0.2.0"; // Current app version
 
 export const DAILY_FILE_UPLOAD_LIMIT = 5;
 
 export const NOTIFICATION_AUDIO_PATH = "/audio/yapsnotify.mp3";
 
 export const MAX_FILE_SIZE = 5 * 1024 * 1024 // 5MB
+
+export const DATABASE_NAME = "wei-database";
+
+export const DATABASE_VERSION = 3;